Added index.html file
Build a hosted web app that calculates the air distance between two locations.

The user will be able to input the longitude and latitude for each location and then press a button to calculate the air distance. Upon pressing the button, the correct air distance should be displayed on the page.

You can earn bonus points by integrating a free map API to let the user input an address that gives the user the longitude and latitude of each location.

There should be two inputs labeled as "Point A" and "Point B" where the user will input the coordinates for the respective points. The coordinates are expected to be latitude and longitude (both in decimal form) separated by a comma.

To get the distance the user needs to click on a "Calculate" button and the results should be displayed in kilometers, in the following format: "Distance: 2.84km"